backend/requirements.txt
channels
channels_redis
daphne
django
djangorestframework
django-cors-headers
psycopg2
